Template/source/requirements.txt
troposphere
cfn-flip
GitPython
python-gitlab
requests
awacs